requirements.txt
watchdog==3.0.0
pillow==10.0.0
piexif==1.1.3
ffmpeg-python==0.2.0